moved the logic of the login, from validation to the submit function. Added a response in case the game cant get the matches from the API. Corrected the way the state of the game was changed.
