+++ react/tic-tac-toe/src/app/screens/LogIn/components/RegisterForm/index.js
@@ -21,8 +21,6 @@ class RegisterForm extends Component {
 
 RegisterForm = reduxForm({
   form: 'register',
-  asyncValidate,
-  asyncBlurFields: ['name']
 })(RegisterForm);
 
 export default RegisterForm;
