+++ react/tic-tac-toe/src/app/screens/LogIn/validation/index.js
@@ -1,5 +1,3 @@
-import logIn from '../../../../services/LoginService';
-
 export const required = value => 
   value ? undefined : "Value is required";
 
@@ -12,22 +10,3 @@ export const matchesPassword = (value, allValues) =>
   value === allValues.password
     ? undefined
     : 'Passwords must match.';
-
-export const asyncValidate = async values => {
-  const sleep = ms => new Promise(resolve => setTimeout(resolve, ms));
-  await sleep(1000);
-
-  const response = await logIn.logIn({email: values.email, password: values.password});
-  if (response.status === 401) {
-    return Promise.reject({
-      email: 'Unauthorized',
-      password: 'Unauthorized'
-    });
-  }
-  if(!response){
-    window.alert("Could not connect to the server");
-    return Promise.reject("Could not connect to the server");
-  }
-
-  window.localStorage.setItem("token", response.data.token);
-}
